requirements.txt
# ─── Core ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0

# ─── Data ────────────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0

# ─── GCP ─────────────────────────────────────────────────────────────────────
google-cloud-bigquery==3.22.0
google-cloud-bigquery-storage==2.25.0
google-auth==2.29.0

# ─── Modeling ────────────────────────────────────────────────────────────────
scikit-learn==1.5.0
xgboost==2.0.3
prophet==1.1.5
lightweight-mmm==0.1.9
statsmodels==0.14.2

# ─── API ─────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0

# ─── Insights + Deck generation ───────────────────────────────────────────────
anthropic==0.28.0
python-pptx==0.6.23

# ─── Google Slides upload ─────────────────────────────────────────────────────
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ─── Monitoring ──────────────────────────────────────────────────────────────
scipy==1.13.0

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
